pyproject.toml
#
# terraform-provider-tofusoup/pyproject.toml
#
# Terraform provider for TofuSoup registry queries and state inspection
#

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terraform-provider-tofusoup"
dynamic = ["version"]
description = "Terraform provider for TofuSoup registry queries and state inspection"
requires-python = ">=3.11"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Tim Perkins", email = "code@tim.life" }
]
maintainers = [
    { name = "provide.io", email = "code@provide.io" }
]
keywords = ["terraform", "provider", "tofusoup", "opentofu", "registry", "state"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]
dependencies = [
    "pyvider",
    "tofusoup",
    "flavorpack",
    "provide-foundation[all]",
]

[project.urls]
Homepage = "https://foundry.provide.io/terraform-provider-tofusoup/"
Documentation = "https://foundry.provide.io/terraform-provider-tofusoup/"
Repository = "https://foundry.provide.io/terraform-provider-tofusoup/repo"
Issues = "https://foundry.provide.io/terraform-provider-tofusoup/issues"

[project.scripts]
terraform-provider-tofusoup = "pyvider.cli:main"

[project.entry-points.pyvider]
terraform-provider-tofusoup = "tofusoup.tf.components"

[dependency-groups]
dev = [
    "provide-testkit[all]",
    "plating",
]
docs = [
    "provide-testkit[docs]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.ruff]
line-length = 111
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "B", "BLE", "C90", "PIE", "PLR", "S", "SIM", "PTH", "RUF"]
ignore = [
    "ANN401",   # any-type (too restrictive)
    "BLE001",   # blind-except (too restrictive for cleanup patterns)
    "B008",     # function-call-in-default-argument (Click uses this pattern)
    "E501",     # line-too-long (handled by formatter)
    "S101",     # assert (valid in tests)
    "S104",     # binding-to-all-interfaces (intentional for servers)
    "S108",     # hardcoded-temp-file (/tmp is fine for our use cases)
    "S310",     # suspicious-url-open-for-protocol (URLs are controlled)
    "S603",     # subprocess-without-shell-equals-true (trusted input)
    "S607",     # start-process-with-partial-path (trusted executables)
    "PLR0911",  # too-many-return-statements (refactor later)
    "PLR0912",  # too-many-branches (refactor later)
    "PLR0913",  # too-many-arguments (common in config classes)
    "PLR0915",  # too-many-statements (refactor later)
    "PLR2004",  # magic-value-comparison (too noisy)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN204", "B017", "C901", "E402", "PLR0124", "PLR1714", "RUF043", "S102", "S105", "S106", "S110", "S307", "S311", "S324", "S602", "SIM105", "SIM108", "SIM117"]
# Complex state inspection data sources
"src/tofusoup/tf/components/data_sources/state_info.py" = ["C901"]
"src/tofusoup/tf/components/data_sources/state_outputs.py" = ["C901"]
"src/tofusoup/tf/components/data_sources/state_resources.py" = ["C901"]

[tool.ruff.lint.isort]
known-first-party = ["tofusoup", "tests"]
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "pyvider.*",
    "tofusoup.*",
    "flavorpack.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ""
# Parallel execution: pytest -n <workers> (range: 2-16, recommend: CPU cores or 8, max: 16 to avoid xdist hang)

[tool.coverage.run]
branch = true
source = ["."]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_covered = true
precision = 2
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "pass",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[pyvider]
component_packages = ["tofusoup.tf.components"]

[tool.plating]
provider_name = "tofusoup"
output_dir = "docs"

[tool.flavor]
entry_point = "pyvider.cli:main"
output_path = "dist/terraform-provider-tofusoup"
platforms = ["darwin_arm64", "linux_amd64", "linux_arm64", "windows_amd64"]

[tool.bandit]
exclude_dirs = [".venv", "tests", "workenv"]

################################################################################
# Security Scanning Allowlist
################################################################################

[tool.security]
# Paths to exclude from secret scanning (TruffleHog, GitGuardian, Gitleaks)
# Run: we security generate (via wrknv) to regenerate tool configs
description = "Test signing keys - intentionally committed"
allowed_paths = [
    # Test signing keys for provider development
    "keys/**",
]
